+++ metatensor-torch/CHANGELOG.md
@@ -30,6 +30,8 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 - `LabelsView` has been removed, and with it the following functions:
   `Labels.is_view()`, `Labels.to_owned()`, `Labels.view()`, and
   `Labels.__getitem__(list[str])`. We recomend using `Labels.column()` instead to access the values of individual dimensions of Labels.
+- Removed `TensorMap.blocks_matching`, the same behavior can be achieved
+  with `TensorMap.keys.select(...)`.
 
 ## [Version 0.8.5](https://github.com/metatensor/metatensor/releases/tag/metatensor-torch-v0.8.5) - 2026-03-27